devel: see ChangeLog
git-svn-id: https://svn.code.sf.net/p/xsnow/code/trunk@154 79837d8a-40e5-47fd-869e-05d31bbd7fe9

+++ xsnow-devel/src/dsimple.c
@@ -355,7 +355,7 @@ Window Window_With_Name(
    if (!XQueryTree(dpy, top, &dummy, &dummy, &children, &nchildren))
       return(0);
 
-   for (i=0; i<nchildren; i++) {
+   for (i=0; (unsigned int)i<nchildren; i++) {
       w = Window_With_Name(dpy, children[i], name);
       if (w)
 	 break;
